+++ src/Builder/ExpressionInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Solarium\Builder;
+
+/**
+ * Expression Interface.
+ *
+ * @author wicliff <wicliff.wolda@gmail.com>
+ */
+interface ExpressionInterface
+{
+    /**
+     * @param \Solarium\Builder\AbstractExpressionVisitor $visitor
+     *
+     * @return mixed
+     */
+    public function visit(AbstractExpressionVisitor $visitor);
+}